remove old folder and fix dialog test
