+++ packages/components/src/Overlay/Overlay.tsx
@@ -1,29 +1,58 @@
-import React, { ReactNode } from 'react';
-import { OverlayContainer } from '@react-aria/overlays';
+import React, { useRef } from 'react';
+import { Transition } from 'react-transition-group';
 
-import { ComponentProps } from '@marigold/types';
+import {
+  Overlay as ReactAriaOverlay,
+  OverlayProps as ReactAriaOverlayProps,
+} from '@react-aria/overlays';
 
-import { Box } from '../Box';
-
-export interface OverlayProps extends ComponentProps<'div'> {
-  children: ReactNode;
-  open?: boolean;
-  container?: HTMLElement;
+export interface OverlayProps {
+  open: boolean;
+  children: ReactAriaOverlayProps['children'];
+  container?: ReactAriaOverlayProps['portalContainer'];
 }
 
-export const Overlay = ({
-  children,
-  open = false,
-  container,
-  ...props
-}: OverlayProps) => {
-  if (!open) {
+const duration = 300;
+
+const defaultStyle = {
+  transition: `opacity ${duration}ms ease-in-out`,
+  opacity: 0,
+};
+
+const transitionStyles = {
+  entering: { opacity: 1 },
+  entered: { opacity: 1 },
+  exiting: { opacity: 0 },
+  exited: { opacity: 0 },
+  unmounted: { opacity: 0 },
+};
+
+export const Overlay = ({ children, container, open }: OverlayProps) => {
+  const nodeRef = useRef(null);
+
+  // Don't un-render the overlay while it's transitioning out.
+  let mountOverlay = open;
+  if (!mountOverlay) {
+    // Don't bother showing anything if we don't have to.
     return null;
   }
 
   return (
-    <OverlayContainer portalContainer={container}>
-      <Box {...props}>{children}</Box>
-    </OverlayContainer>
+    <ReactAriaOverlay portalContainer={container}>
+      <Transition nodeRef={nodeRef} timeout={{ enter: 0, exit: 350 }} in={open}>
+        {state => (
+          <div
+            ref={nodeRef}
+            data-testid="overlay"
+            style={{
+              ...defaultStyle,
+              ...transitionStyles[state],
+            }}
+          >
+            {children}
+          </div>
+        )}
+      </Transition>
+    </ReactAriaOverlay>
   );
 };
