+++ packages/components/src/Overlay/Popover.tsx
@@ -1,74 +1,86 @@
-import React, { forwardRef, ReactNode, useRef } from 'react';
-import { AriaOverlayProps, useModal, useOverlay } from '@react-aria/overlays';
-import { mergeProps } from '@react-aria/utils';
-
-import { Box } from '../Box';
+import React, { forwardRef, ReactNode, RefObject } from 'react';
+import {
+  AriaPopoverProps,
+  DismissButton,
+  usePopover,
+} from '@react-aria/overlays';
+import { OverlayTriggerState } from '@react-stately/overlays';
 import { Overlay } from './Overlay';
+import { Underlay } from './Underlay';
+import { useObjectRef } from '@react-aria/utils';
+import { FocusScope } from '@react-aria/focus';
 
 export interface PopoverProps
-  extends Omit<
-    AriaOverlayProps,
-    'isOpen' | 'isDismissable' | 'isKeyboardDismissDisabled'
-  > {
-  children?: ReactNode;
-  open?: boolean;
-  dismissable?: boolean;
-  keyboardDismissDisabled?: boolean;
-
-  /**
-   * Adjust size of the popover. This is used to make the popover
-   * at least the same width as its anchor element.
-   */
-  minWidth?: number | string;
+  extends Pick<AriaPopoverProps, 'triggerRef' | 'scrollRef' | 'isNonModal'> {
+  keyboardDismissDisabled?: AriaPopoverProps['isKeyboardDismissDisabled'];
+  state: OverlayTriggerState;
+  children: ReactNode;
 }
 
-export const Popover = forwardRef(
+//   /**
+//    * Adjust size of the popover. This is used to make the popover
+//    * at least the same width as its anchor element.
+//    */
+//   minWidth?: number | string;
+// }
+
+interface PopoverWrapperProps extends PopoverProps {}
+
+export const Popover = forwardRef<HTMLDivElement, PopoverProps>(
+  (props, ref) => {
+    const popoverRef = useObjectRef(ref);
+    let { children, state, ...otherProps } = props;
+    return (
+      <Overlay open={state.isOpen} {...otherProps}>
+        <PopoverWrapper ref={popoverRef} {...props}>
+          {children}
+        </PopoverWrapper>
+      </Overlay>
+    );
+  }
+);
+
+const PopoverWrapper = forwardRef(
   (
     {
       children,
-      open,
-      dismissable,
+      isNonModal,
+      state,
       keyboardDismissDisabled,
-      shouldCloseOnBlur,
-      minWidth = 0,
       ...props
-    }: PopoverProps,
-    ref
+    }: PopoverWrapperProps,
+    ref: RefObject<HTMLDivElement>
   ) => {
-    // FIXME: Why can't we use `useObjectRef` here?
-    const fallbackRef = useRef(null);
-    const popoverRef = ref || fallbackRef;
-    const { overlayProps } = useOverlay(
+    const { popoverProps, underlayProps } = usePopover(
       {
-        isOpen: open,
-        isDismissable: dismissable,
-        isKeyboardDismissDisabled: keyboardDismissDisabled,
-        shouldCloseOnBlur,
         ...props,
+        isNonModal,
+        isKeyboardDismissDisabled: keyboardDismissDisabled,
+        popoverRef: ref,
+        placement: 'bottom left',
       },
-      popoverRef as any
+      state
     );
 
-    /**
-     * Hide content outside the container from screen readers.
-     */
-    const { modalProps } = useModal({});
-
-    /**
-     * Fit size to anchor element if children contents is smaller.
-     */
-    const style = { minWidth };
-
     return (
-      <Overlay open={open}>
-        <Box
-          ref={popoverRef}
+      <FocusScope restoreFocus>
+        {!isNonModal && <Underlay {...underlayProps} />}
+        <div
+          {...popoverProps}
+          style={{
+            ...popoverProps.style,
+            minWidth: props.triggerRef.current
+              ? (props.triggerRef.current as HTMLElement).offsetWidth
+              : undefined,
+          }}
+          ref={ref}
           role="presentation"
-          {...mergeProps(props, overlayProps, modalProps, style)}
         >
+          {!isNonModal && <DismissButton onDismiss={state.close} />}
           {children}
-        </Box>
-      </Overlay>
+          <DismissButton onDismiss={state.close} />
+        </div>
+      </FocusScope>
     );
   }
 );
