+++ packages/components/src/Overlay/index.ts
@@ -1,3 +1,4 @@
+export * from './Modal';
 export * from './Overlay';
 export * from './Popover';
 export * from './Underlay';
