+++ packages/components/src/Overlay/overlay.test.tsx
@@ -1,7 +1,9 @@
 /* eslint-disable testing-library/no-node-access */
-import React from 'react';
-import { render, screen } from '@testing-library/react';
+import React, { forwardRef } from 'react';
+import { render, screen, waitFor } from '@testing-library/react';
 import { OverlayProvider } from '@react-aria/overlays';
+import { useOverlayTriggerState } from '@react-stately/overlays';
+import { useObjectRef } from '@react-aria/utils';
 import { ThemeProvider } from '@marigold/system';
 import { Overlay } from './Overlay';
 import { Popover } from './Popover';
@@ -52,7 +54,7 @@ test('renders open overlay', () => {
 test('overlay is per default closed', () => {
   render(
     <OverlayProvider data-testid="overlay">
-      <Overlay>
+      <Overlay open={false}>
         <div>something</div>
       </Overlay>
     </OverlayProvider>
@@ -92,57 +94,67 @@ test('overlay has container', () => {
 
 // Popover tests
 // ---------------
-test('renders open popover', () => {
-  render(
-    <OverlayProvider>
-      <Popover data-testid="popover" open>
-        <div>something</div>
-      </Popover>
-    </OverlayProvider>
-  );
+interface TestPopoverProps {
+  open: boolean;
+}
+const TestPopover = forwardRef<HTMLDivElement, TestPopoverProps>(
+  ({ open }, ref) => {
+    const popoverRef = useObjectRef(ref);
+    const state = useOverlayTriggerState({ isOpen: open });
 
-  const popover = screen.getByTestId('popover');
+    return (
+      <OverlayProvider>
+        <Popover triggerRef={popoverRef} state={state}>
+          <div>something</div>
+        </Popover>
+      </OverlayProvider>
+    );
+  }
+);
+
+test('renders open popover', () => {
+  render(<TestPopover open={true} />);
+
+  const popover = screen.getByRole('presentation');
   expect(popover).toBeInTheDocument();
 });
 
 test('popover is per default closed', () => {
-  render(
-    <OverlayProvider data-testid="popover">
-      <Popover>
-        <div>something</div>
-      </Popover>
-    </OverlayProvider>
-  );
+  render(<TestPopover open={false} />);
 
-  const popover = screen.getByTestId('popover').firstChild;
+  const popover = screen.queryByRole('presentation');
   expect(popover).not.toBeInTheDocument();
 });
 
-test('forwards ref', () => {
-  const ref = React.createRef<HTMLInputElement>();
-  render(
-    <OverlayProvider>
-      <Popover data-testid="popover" open ref={ref}>
-        <div>something</div>
-      </Popover>
-    </OverlayProvider>
-  );
+test('forwards ref', async () => {
+  const ref = React.createRef<HTMLDivElement>();
 
-  expect(ref.current).toBeInstanceOf(HTMLDivElement);
+  render(<TestPopover open={true} ref={ref} />);
+
+  waitFor(() => {
+    expect(ref.current).toBeInstanceOf(HTMLDivElement);
+  });
 });
 
 test('popover has children', () => {
-  render(
-    <OverlayProvider>
-      <Overlay data-testid="overlay" open>
-        <div>something</div>
-      </Overlay>
-    </OverlayProvider>
-  );
+  render(<TestPopover open={true} />);
 
-  const popover = screen.getByTestId('overlay');
+  const popover = screen.getByRole('presentation');
   expect(popover).toBeInTheDocument();
   expect(popover.firstChild).toBeInTheDocument();
+});
+
+test('minWidth is set', async () => {
+  const ref = React.createRef<HTMLDivElement>();
+  render(<TestPopover open={true} ref={ref} />);
+
+  const popover = screen.getByRole('presentation');
+
+  waitFor(() => {
+    expect(popover).toHaveStyle(
+      `min-width: ${(ref.current as HTMLElement).offsetWidth}`
+    );
+  });
 });
 
 // Underlay tests
